DDA-17: Create add room tool
