+++ src/react/features/connector/tools/buttons.js
@@ -1,0 +1,11 @@
+export const ADD_ROOM = 'ADD_ROOM'
+
+export default [
+  {
+    id: ADD_ROOM,
+    text: 'Add room',
+    data: {
+      icon: 'plus',
+    },
+  },
+]
